requirements.txt
# Dependencias base para ejecutar el cotizador y el backend opcional
flask>=2.2,<3.0
python-docx>=0.8.11
docx2pdf>=0.1.8
